+++ lib/src/config/config.dart
@@ -14,6 +14,12 @@ class NotificationToolsConfigApp {
   /// Outillage
   ///
 
+  // nom de l'application
+  static String appName = "Ricochets App";
+
+  // copyright de l'application
+  static String appCopyrightName = "Ricochets Développement";
+
   // gestion des notifications
   static final mbNotifications notifications = mbNotifications(
     'mipmap/ic_launcher',